Spajanje grane "filip" na granu "filip+aleksandar"...
11) Dodata klasa ComponentMover, preuzeta sa interneta. Ima dosta
funkcionalnosti, od kojih je najbitnija pomijeranje komponenti unutar
panela (drag & drop).

Iz dokumentacije: This class allows you to move a Component by using a
mouse. The Component moved can be a high level Window (ie. Window,
Frame, Dialog) in which case the Window is moved within the desktop. Or
the Component can belong to a Container in which case the Component is
moved within the Container. When moving a Window, the listener can be
added to a child Component of the Window. In this case attempting to
move the child will result in the Window moving. For example, you might
create a custom "Title Bar" for an undecorated Window and moving of the
Window is accomplished by moving the title bar only. Multiple components
can be registered as "window movers". Components can be registered when
the class is created. Additional components can be added at any time
using the registerComponent() method.